fix(Dapr): Fix the error in obtaining dapr process, causing dapr to be restarted (#67)
* fix(Dapr): Fix the error in obtaining dapr process, causing dapr to be restarted

* fix(dapr): Fix dapr service being restarted repeatedly

+++ src/Development/Masa.Utils.Development.Dapr/Process/SystemProcess.cs
@@ -20,8 +20,7 @@ public class SystemProcess : IProcess
 
     public void Kill()
     {
-        if (IsRun)
-            _process.Kill();
+        if (IsRun) _process.Kill();
     }
 
     public bool Start() => _process.Start();
